chore: move should ignore node to global util
